fix: extract MessengerSetupCard to prevent state reset on re-render
Root cause: MessengerSetupCard was defined INSIDE the OnboardingWizard
component function. React treats inline component definitions as new
component types on every render, causing unmount/remount cycles that
reset all internal state (status, botLink, qrCode, etc).

This caused:
- Telegram setup stuck in 'WAITING' even after server was online
- Rapid flipping between states as the card kept getting recreated
- Setup API being called repeatedly in a loop
- Unreadable flashing text on mobile

Fix: Extract MessengerSetupCard to its own module-level file
(messenger-setup-card.tsx) and import it. React now sees it as a
stable component type that persists across parent re-renders.
